requirements.txt
ipykernel
datasets
pillow
matplotlib
seaborn
pandas
numpy
requests
torchvision
transformers
scikit-learn
sentencepiece
sacremoses
streamlit
git+https://github.com/openai/CLIP.git
sentence_transformers
tensorflow
tf-keras